+++ src/app/NoteUseCaseFactory.java
@@ -14,10 +14,10 @@ import use_case.note.save.SaveNoteInputBoundary;
 import use_case.note.save.SaveNoteInteractor;
 import use_case.note.save.SaveNoteOutputBoundary;
 
-import use_case.quiz.DisplayQuizzesInputBoundary;
-import use_case.quiz.DisplayQuizzesInteractor;
-import use_case.quiz.DisplayQuizzesOutputBoundary;
 import use_case.quiz.QuizDataAccessInterface;
+import use_case.quiz.display_quizzes.DisplayQuizzesInputBoundary;
+import use_case.quiz.display_quizzes.DisplayQuizzesInteractor;
+import use_case.quiz.display_quizzes.DisplayQuizzesOutputBoundary;
 import view.NoteView;
 
 public class NoteUseCaseFactory {
