+++ src/interface_adapter/quiz/display_quizzes/DisplayQuizzesController.java
@@ -1,6 +1,6 @@
 package interface_adapter.quiz.display_quizzes;
 
-import use_case.quiz.DisplayQuizzesInputBoundary;
+import use_case.quiz.display_quizzes.DisplayQuizzesInputBoundary;
 
 public class DisplayQuizzesController {
