--- src/use_case/quiz/DisplayQuizzesInputBoundary.java
@@ -1,5 +1,0 @@
-package use_case.quiz;
-
-public interface DisplayQuizzesInputBoundary {
-    void execute();
-}
