--- src/use_case/quiz/DisplayQuizzesInputData.java
@@ -1,18 +1,0 @@
-package use_case.quiz;
-
-public class DisplayQuizzesInputData {
-    //TODO: this class is not needed, we can delete it later
-//    private String title;
-//
-//    public DisplayQuizzesInputData(String title){
-//        this.title = title;
-//    }
-//
-//    public String getTitle() {
-//        return this.title;
-//    }
-//
-//    public void setTitle(String title) {
-//        this.title = title;
-//    }
-}
